Review Stars
The reputation stars are now working. They correspond to the users reputation in the database.

I also changed the way that the database info is inputted in the h1.

Instead of using made up varialbes, i create a new instance of the UserProfiles class and add the info from database to the attributes of that class, and then i use it via @CurrentUser.Firstname etc.